requirements.txt
pycairo==1.25.1
pygame==2.5.2
